[DDW-781] Blank Screen Fix - Default Selection. (#1488)
* [DDW-781] Blank Screen Fix - Default Selection.

* [DDW-711] Replaces button No by Cancel

* [DDW-781] Fixes buttons focus

* DDW-781 CHANGELOG
